split prompt header classes into separate files by their type
